Customer objects are put into hazelcast and then fetched to print to the console
HazelcastDaoUtil utility class is created for performing hazelcast operations with hazelcastCustomerDaoImpl class

HazelcastCustomerDaoImpl is created for realizing data access and storage operations with Hazelcast

CustomerDao interface is created for seperating application logic from data access and storage concern

CustomerCompactSerializer class is created for using native Hazelcast serializer instead of default java serializer when we do put and get operations on the IMap with our Customer objects

Customer class is implemented from Serializable class to use default java serializer when we put Customer object in a Hazelcast IMap as a serialized object. Parameterized constructor is added for creating Customer object with values that are read with CompactReader class. And toString method is added for printing Customer object.

Constants Enum is created for important string info